+++ src/main/java/org/folio/qm/exception/MappingMetadataException.java
@@ -1,0 +1,16 @@
+package org.folio.qm.exception;
+
+public class MappingMetadataException extends RuntimeException {
+
+  public MappingMetadataException(String message) {
+    super(message);
+  }
+
+  public MappingMetadataException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public MappingMetadataException(Throwable cause) {
+    super(cause);
+  }
+}