+++ NEWS.md
@@ -5,6 +5,11 @@
 ### New APIs versions
 * Provides `API_NAME vX.Y`
 * Changed to optional: `instance-authority-links`, `instance-authority-linking-rules`, `instance-authority-links-suggestions`
+* Optional `mapping-metadata-provider v1.1`
+* Optional `authority-storage v2.0`
+* Optional `instance-storage v11.2`
+* Optional `holdings-storage v8.8`
+* Optional `instance-preceding-succeeding-titles v0.1`
 
 ### Features
 * Add “allowedValues” to the MARC authority 008 spec response ([MODQM-455](https://folio-org.atlassian.net/browse/MODQM-455))
@@ -12,6 +17,7 @@
 * Convert subfield to lowercase during quick-marc to source record convertion ([MODQM-477](https://folio-org.atlassian.net/browse/MODQM-477))
 * Use unique consumer group per instance for specification-storage.specification.updated topic ([MODQM-483](https://folio-org.atlassian.net/browse/MODQM-483))
 * Ignore empty subfields during conversion from DTO ([MODQM-445](https://folio-org.atlassian.net/browse/MODQM-445))
+* Implement MARC Record Update Flow ([MODQM-490](https://folio-org.atlassian.net/browse/MODQM-490))
 
 ### Bug fixes
 * Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
@@ -26,6 +32,8 @@
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`
 * Add `LIB_NAME VERSION`
 * Remove `LIB_NAME`
+* Add `data-import-processing-core v4.5.0-SNAPSHOT`
+* Add `mod-inventory-storage-dto v30.0.0-SNAPSHOT`
 
 ---